Move petition preview to legacy components
